+++ prices/src/main/java/com/challenge/prices/domain/PriceRepository.java
@@ -1,10 +1,10 @@
-package com.challenge.prices.domain.port.out;
+package com.challenge.prices.domain;
 
 import com.challenge.prices.domain.models.Price;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface GetPricePort {
+public interface PriceRepository {
     Optional<Price> getPrice(Long productId, Long brandId, LocalDateTime applicationDate);
 }
