+++ prices/src/main/java/com/challenge/prices/domain/PriceService.java
@@ -1,10 +1,10 @@
-package com.challenge.prices.domain.port.in;
+package com.challenge.prices.domain;
 
-import com.challenge.prices.util.dto.PriceDTO;
+import com.challenge.prices.domain.models.PriceDTO;
 
 import java.time.LocalDateTime;
 import java.util.Optional;
 
-public interface PriceServiceUseCase {
+public interface PriceService {
     Optional<PriceDTO> getPrice(Long productId, Long brandId, LocalDateTime applicationDate);
 }
