+++ prices/src/main/java/com/challenge/prices/infrastructure/repository/PriceRepositoryAdapter.java
@@ -1,7 +1,7 @@
-package com.challenge.prices.infrastructure.outbound.database;
+package com.challenge.prices.infrastructure.repository;
 
 import com.challenge.prices.domain.models.Price;
-import com.challenge.prices.domain.port.out.GetPricePort;
+import com.challenge.prices.domain.PriceRepository;
 import org.springframework.stereotype.Component;
 import lombok.RequiredArgsConstructor;
 
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 @Component
 @RequiredArgsConstructor
-public class PriceRepositoryAdapter implements GetPricePort {
+public class PriceRepositoryAdapter implements PriceRepository {
 
     private final JPAPriceRepository jpaPriceRepository;
